Start on building out animator
When building out the animator, I'm realizing
that I'm going to be rebuilding a lot of what
CSS animations are. This might be great for a
game like asteroids, but with a well laid
out grid, it might just not make sense.
I've decided, after a lot of deliberation,
to bring in music-to-us and combine it with
this project. It's more or less the same project
anyways, so this'll be a good combination.
The only thing to reconcile will be tracks and
multiplayer.