Remove call to npm run clean
For some reason in this project it was pointing
to a `rimraf` command, which I dno't even know
what that's supposed to be doing. It really caught me by surprise.

I'm investigating b/c I made a commit in master, ran nr release,
and it undid my changes. They're in the tree, but they got
reverted somehow.